refactor: split project board script into src and test folders
