+++ .github/scripts/project-board/test/snapshot.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import {test} from "node:test";
-import {buildSnapshot, pickProjectItem, type PrNode} from "./snapshot.ts";
+import {buildSnapshot, pickProjectItem, type PrNode} from "../src/snapshot.ts";
 
 function prNode(overrides: Partial<PrNode>): PrNode {
   return {
