+++ .github/scripts/project-board/tsconfig.json
@@ -10,5 +10,5 @@
     "skipLibCheck": true,
     "types": ["node"]
   },
-  "include": ["*.ts"]
+  "include": ["src/*.ts", "test/*.ts"]
 }
